Update search repositories

+++ src/main/java/com/jvm_bloggers/entities/blog_post/BlogPostTextSearchRepository.java
@@ -6,6 +6,6 @@ public interface BlogPostTextSearchRepository {
 
   List<BlogPost> findApprovedPostsByTagOrTitle(String searchPhrase, int page, int pageSize);
 
-  int countApprovedPostsByTagOrTitle(String  searchPhrase);
+  long countApprovedPostsByTagOrTitle(String  searchPhrase);
 
 }
